requirements.txt
psycopg2-binary==2.9.11
dotenv==0.9.0
python-dotenv==1.2.1
matplotlib==3.10.8
numpy==2.2.6
typing==3.7.4.3